requirements.txt
setuptools
flask
pytz
mysql-connector-python
PyJWT
redis
rq
boto3
